+++ lib/react-xjst/adapter.js
