+++ lib/react-xjst/app.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const React = require('react');
-const ReactDomServer = require('react-dom/server');
 const provide = require('./adapter');
 
 class App extends React.Component {
@@ -45,4 +44,4 @@ class Footer extends React.Component {
   }
 }
 
-module.exports = childrenNum => () => ReactDomServer.renderToString(React.createElement(App, { childrenNum }));
+module.exports = App;
