+++ lib/react-xjst/assets/example.ddsl.js
