+++ lib/react-xjst/templates/templates.bemhtml.js
