+++ lib/react-xjst/templates/templates.deps.js
