+++ renderToString/index.js
@@ -5,10 +5,9 @@ const join = require('path').join;
 
 const CHILDREN_NUM = 200;
 const templates = fs.readFileSync(join(
-  process.cwd(), 'renderToString', 'react-xjst', 'templates', 'templates.bemhtml.js'),
+  process.cwd(), 'lib', 'react-xjst', 'templates', 'templates.bemhtml.js'),
   'utf-8'
 );
-console.log(require('./react')(CHILDREN_NUM)());
 
 module.exports = {
   'react': require('./react')(CHILDREN_NUM),
