+++ renderToString/react-xjst.js
@@ -2,6 +2,6 @@
 
 const React = require('react');
 const ReactDomServer = require('react-dom/server');
-const App = require('../lib/react/app');
+const App = require('../lib/react-xjst/app');
 
 module.exports = childrenNum => () => ReactDomServer.renderToString(React.createElement(App, { childrenNum }));
