--- renderToString/react-xjst/templates/templates.deps.js
@@ -1,4 +1,0 @@
-({
-  mustDeps: [],
-  shouldDeps: []
-});
